more warnings fixes

+++ Engine/EditorUI/Panel/PanelProjectExplorer.cpp
@@ -145,7 +145,7 @@ void PanelProjectExplorer::ShowFilesInExplorer()
 	}
 
 	ImVec2 available_region = ImGui::GetContentRegionAvail();
-	int files_per_line = available_region.x / file_size_width;
+	int files_per_line = static_cast<int>(available_region.x / file_size_width);
 
 	int current_line = 0;
 	int current_file_in_line = 0;
@@ -201,7 +201,7 @@ void PanelProjectExplorer::ShowMetafile(Path* metafile_path, Metafile* metafile,
 		ResourceDragSource(metafile);
 		ProcessResourceMouseInput(metafile_path, metafile);
 
-		ImGui::SetCursorPosX((ImGui::GetWindowWidth() - 0.75 * file_size_width) * 0.5f);
+		ImGui::SetCursorPosX((ImGui::GetWindowWidth() - 0.75f * file_size_width) * 0.5f);
 		ShowMetafileIcon(metafile);
 		if (is_model)
 		{
@@ -227,8 +227,8 @@ void PanelProjectExplorer::ShowMetafile(Path* metafile_path, Metafile* metafile,
 			}
 			else
 			{
-				int character_width = text_width / filename.length();
-				int string_position_wrap = file_size_width / character_width - 5;
+				float character_width = text_width / filename.length();
+				size_t string_position_wrap = static_cast<size_t>(file_size_width / character_width - 5);
 				assert(string_position_wrap < filename.length());
 				std::string wrapped_filename = filename.substr(0, string_position_wrap + 3) + "\n" + filename.substr(string_position_wrap, filename.size());
 				ImGui::Text(wrapped_filename.c_str());
@@ -247,7 +247,7 @@ void PanelProjectExplorer::ShowMetafileIcon(Metafile * metafile)
 
 	if(metafile->resource_type == ResourceType::TEXTURE)
 	{
-		ImGui::Image((void *)GetResourcePreviewImage(metafile->uuid), ImVec2(0.75*file_size_width, 0.75*file_size_width));
+		ImGui::Image((void *)GetResourcePreviewImage(metafile->uuid), ImVec2(0.75f*file_size_width, 0.75f*file_size_width));
 	}
 	else
 	{
@@ -297,10 +297,10 @@ void PanelProjectExplorer::ShowMetafileIcon(Metafile * metafile)
 
 		ImGui::PushItemFlag(ImGuiItemFlags_Disabled, true);
 
-		static ImVec4 color = ImVec4(0, 0.4, 0.6, 1);
+		static ImVec4 color = ImVec4(0.f, 0.4f, 0.6f, 1.f);
 		ImGui::PushStyleColor(ImGuiCol_Button, color);
 
-		ImGui::Button(icon.c_str(),ImVec2(0.75*file_size_width, 0.75*file_size_width));
+		ImGui::Button(icon.c_str(),ImVec2(0.75f*file_size_width, 0.75f*file_size_width));
 		ImGui::PopStyleColor(1);
 		ImGui::PopItemFlag();
 	}
